Se corrige la variable del método calcularCosto

+++ Interfaz.java
@@ -1,4 +1,4 @@
 public interface Interfaz {
-    int calcularCosto();
+    double calcularCosto();
     String validarEntrega();
 }
